+++ V2rayNG/app/src/main/java/com/v2ray/ang/core/root/RootManager.kt
@@ -1,4 +1,4 @@
-package com.v2ray.ang.handler
+package com.v2ray.ang.core.root
 
 import com.v2ray.ang.AppConfig
 import com.v2ray.ang.util.LogUtil
